refactor(docs): split cli_entrypoint into per-command modules
cli_entrypoint.py was 1249 lines with every command, render helper,
and mirror glue in one file. Broken out into stacklets/docs/bot/cli/
as one module per command plus three shared `_`-prefixed modules.

  cli/_shared.py      err + parse_doc_id + dry-run flag set
  cli/_dry_run.py     DryRunPaperless wrapper
  cli/_mirror.py      mirror bootstrap + publish_enriched + publish_current_state
  cli/show.py         `stack docs show`
  cli/classify.py     `stack docs classify`     (delegates to reprocess.run_one)
  cli/reformat.py     `stack docs reformat`
  cli/reprocess.py    `stack docs reprocess`    (also exports run_one)
  cli/mirror.py       `stack docs mirror`
  cli/tags.py         `stack docs tags {merge,prune,delete}`

cli_entrypoint.py drops to ~100 lines of dispatch. No behavior change:
all 169 unit tests pass, smoke-tested classify/reformat/reprocess/show
against the live LLM with matching output.

Also fixes two pre-existing failures in test_docs_seed.py — expectations
drifted after taxonomy.toml renamed `Steuern`→`Steuer`, `Taxes`→`Tax`,
`Utilities`→`Utility`. Test file comment still said "taxonomy.yaml"
(toml now), corrected.
